feat(#21): moves duplicate code into an abstract base class
